Final version
Finalized version.  Made classes for "diet" and "habitat".  Gave "has a" relationship to the Animal class that allows it to pull a description for Diet and Habitat of the animal.

Made changes to the output of the animal sound to make more sense within the context of overall output.

Added more notes to showcase important details about the code and highlight use of polymorphism.